example: update FizzBuzz to run on 3 runtimes
The source and Fizz operator run on the runtime "foo".

The first Buzz and Receiver run on "bar".
The second Buzz2 and Receiver2 run on "baz" and Buzz2 outputs "Buzz2" instead of
"Buzz".
The third Buzz3 and Receiver3 run on "foo" (the same runtime as the source and
Fizz) and Buzz3 outputs "Buzz3".

This example showcases several behaviors:
- only one publisher is created although we are sending the data to 2 runtimes,
- a link + a subscriber are created (the link for Buzz3 which is on the same
runtime as Fizz and the subscriber for Buzz and Buzz2).