refacto: simplified links declaration in YAML
The new syntax to declare links is:

```yaml
links:
  - from: component_id.output_id
    to: component_id.input_id
```

To achieve this change, customs serializer / deserializer and new descriptors
were implemented.

The core of the changes are in `src/model/link.rs`. The serializers simply take
a string as input and split it at the '.' character.

All the examples were updated accordingly.