feat: add Loop support
A Loop is written as following:

```yaml
loops:
  - ingress: A
    egress: B
    feedback_port: feedback
    port_type: int
    is_infinite: false
```

The `ingress` is the starting point of a Loop. The `egress` is where it can
end. Each time a message goes through a Loop, Zenoh Flow either:
- attaches a new `LoopContext` to it,
- updates the already attached `LoopContext`.

For a given message, the `LoopContext` provides:
- the timestamp at which it entered the Loop,
- the timestamp at which the current iteration started,
- the duration of the last iteration,
- the iteration count (if the Loop was declared as finite).

A Loop between Operators can only exist if the following constraints are respected:
- the Ingress (starting point of the Loop) has a single output,
- the Egress (ending point) has a single input,
- no path between both Operators exist (excluding other Loops),
- an Operator can only be Ingress/Egress for a single Loop.

The dataflow validator was updated accordingly. In particular, the internal
structures now take the `NodeKind` into consideration (to be able to enforce the
constraints on Loops) and keeps track of the nodes that are Ingress / Egress.

---

Additional internal modifications include:
- The `DataFlowRecord` now uses `HashMap` to keep track of the nodes. This was
  motivated by the need to modify them when adding the feedback links of Loops.
- The validator checks that, without Loops, the dataflow is a Directed Acyclic
  Graph.