feat: add global configuration (+ small refacto of mapping)
This commit introduces the possiblity to define a global configuration that all
nodes of the data flow will have.

Values in this configuration can be locally overwritten (in the configuration of
the node).

---

A small refactoring of the mapping was also performed: the custom struct was
replaced by a HashMap, which eases the manipulation.