feat: optionally toggle parts of a data flow using flags
A flag allows user to specify which part of a data flow to activate or
deactivate by declaring a set of nodes.

Nodes that are not present under any flag are activated.

The validity of toggled flags is checked, raising an error if need be. For flags
that are not toggled, only a warning is emitted.