fix: use Fn instead of FnOnce/FnMut in closures
So users can keep a state in the closures they pass to Zenoh-Flow to execute,
the different traits have been updated and expect an `Fn`.

Given these closures are in asynchronous blocks, the only way to mutate any
state they carry is to leverage the interior mutability pattern (i.e. Arc +
Mutex).