feat: new Configuration propagation strategy
With the introduction of composite operators, the way the Configuration was
propagated between nodes needed to be updated.

This commit introduces the updated strategy. In particular, the "hierarchy"
between Configuration is as following:

data flow "inner"
 > data flow "outer"
 > composite "inner"
 > composite "outer"
 > node

"inner" and "outer" only apply to data flow and composite operators
declarations. They respectively refer to the configuration specified at the node
level and at the descriptor level (i.e. the YAML file). Below is a small YAML
file showing the difference.

```yaml
flow: my data flow

configuration:   # outer level
  key1: value1

sources:
  - id: source_1
    uri: file://./source.so
    configuration:  # inner level
      key2: value2
```

This hierarchy allows modifying the configuration of a node directly from the
data flow descriptor.