wip: test implementation of new traits `*Factory` and `Node`
The `*Factory` traits are a way of generating nodes.

The `Node` trait is shared between all kind of nodes. It contains a single
stateful method `iteration` that Zenoh-Flow calls in an infinite loop.

The objective of this test is to assess if these new "clearer" (or so we feel)
traits impact negatively the performance. Our main motivation to test it is that
the current state of Zenoh-Flow v0.4.0 forces users to define asynchronous
closures which have a complex "ugly" (or so we feel) syntax.