feat: InputBuilder and OutputBuilder
Instead of having the methods `take` and `take_raw` on the Inputs and Outputs,
that give, respectively, the Typed and Raw Input/Output, this commit removes
the `take_raw` variant and introduces builders.

The builders can be turned into the Typed or Raw variant using the corresponding
methods: `build_typed` and `build_raw`.