refacto: move records into their own crate
This commit creates the package `zenoh-flow-records`.

The central structure is the `DataFlowRecord`: a data flow that is ready to be
instantiated. All nodes are mapped to a runtime and the required connectors were
added.

The preferred way to create a `DataFlowRecord` is to call the method
`complete_mapping_and_connect` on a `FlattenedDataFlowDescriptor` instance.

This commit was also the opportunity to:
- remove the `runtime` field from all the descriptors / records,
- the remaining `String` were transformed into `Arc<str>`,

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>