refacto(shared-memory): update deserialize + remove `element_size`
Instead of deserializing the size to a `u64` that we then cast into a `usize` we
now attempt to cast it and if that fails we return an error. We apply the same
procedure to the time (casting from `u128` to a `u64`).

The field `element_size` was removed as it is not needed when allocating shared
memory: we first serialize the data and then ask the shared memory manager for
just the amount of memory needed.

The field `total_size` was also renamed to `size`.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>