refacto(descriptors): move uri at root of crate
Parsing a structure from a uri is used outside of the `zenoh-flow-descriptors`
crate it thus makes more sense to expose it at the root of the crate.

Also, the function `try_load_descriptor_from_file` was renamed to
`try_load_from_file` as not only descriptors can be loaded.

Maybe it would even be better to move this function the commons crate?

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>