refacto(runtime): expose Loader in public API
The daemon, plugin and standalone runtime will need to load a configuration for
the loader in order to be able to use nodes implemented in programming languages
other than Rust.

It thus need to be public.

This commit also refactors a bit the internal structure of `LoaderConfig` in
order to automatically obtain a HashMap.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>