feat(runtime): make Zenoh and Zenoh's Shared Memory, features
By making them features, it is then possible to create a zenoh-flow-runtime that
is purely local.

Users that will deploy their nodes on a single machine then won't have to
include Zenoh in their dependencies.

Similarly, if users do not need shared memory (as it could have a hidden extra
cost) then it can be removed.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>