refacto(runtime): `start` and `abort`* all nodes instead of by group
`abort` was previously called `stop` which did not reflect the behavior that is
implemented.

The `start` and `abort` methods now operate globally instead of by group of
nodes, depending on their kind (Sink -> Operator -> Source).
This separation was justified by a desire to not loose any data. In truth, no
message would be lost, as the communication channels are created before the
nodes are started so the messages would stay there.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>