refacto(runtime): make the `Runtime` concurrency-safe
(or at least try to)

Considering that the Runtime could be embedded inside a Daemon that can receive
multiple requests in parallel, concurrency issues could have happened.

These changes aim at mitigating these issues while minimizing the contention on
the libraries and the flows and guaranteeing a consistent view of the flows.

The main modifications were:
- renamed `LoaderConfig` to `Extensions`,
- put the `Loader` inside a `Mutex`,
- put each data flow instance inside an `Arc<Mutex<...>>`,
- put the `HashMap` containing the flows inside a `Mutex`.

Putting each data flow behind an `Arc<Mutex<...>>` allows operating on a flow
without having to lock the entire `HashMap`: a request can clone the `Arc` and
operate on the flow independently of other requests.

The consistent view of the flows is notably ensured by, when instantiating a
flow, inserting an already locked mutex in the `HashMap`. Subsequent operations
on the very same flow would then have to wait for it to have finished.

The standalone runtime was updated accordingly.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>