feat(nodes): add `on_resume` and `on_abort` methods to Node trait
When aborting a data flow, the execution of nodes is interrupted at an `await`
point. The purpose of these methods is to allow users to make their nodes react
to such an interruption correctly.

The modifications we made to the builtin Zenoh source and sink perfectly
illustrate the impact and need for these methods.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>