feat(standalone-daemon): (wip) standalone Zenoh-Flow daemon
The standalone daemon is an executable version of the Zenoh-Flow daemon.

By default it connects to the local Zenoh network as a peer. It allows setting
the identifier of the underlying Zenoh-Flow runtime through a command line
argument.

(WIP) Although it also accepts a configuration through command line parameters,
this parameter is for now ignored.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>