refacto: disable shared-memory feature by default
The current implementation is not as optimised as it deserves to be. In short,
if the shared memory is enabled, then every time Zenoh-Flow will send data
through Zenoh (be it between runtimes on the same host or on different hosts) it
will first attempt to send it through shared memory.

As it is now, sending data through shared memory requires a copy as the data
does not reside inside a shared memory buffer (it must be copied from the
programs memory to the shared memory).

In cases where there is actually a shared memory buffer, this could still lead
to better performance. However, as the current design enables it *everywhere*,
if there is no shared memory then the data is copied twice: (i) from the
programs memory to the shared memory and (ii) from the shared memory to a Zenoh
buffer to be sent on the network.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>