doc(commons): improve documentation of Configuration and Vars
This commit tries to explain how these structures are leveraged by Zenoh-Flow
and focuses in particular and how they are merged.

Users can declare a `Configuration` and `Vars` at several locations when
describing their application. When this happens Zenoh-Flow merges these
declaration in a specific way. This commit tries to clarify the "merge-strategy"
employed.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>