refacto(descriptors): make description section optional
Before this commit, the `description` field for each node was mandatory. It is
now optional for all.

Additionally, when two descriptions are provided (e.g. inside the descriptor
file of a node + when it is included in the data flow descriptor), the "outer"
description will overwrite the "inner" one. See the modifications made in the
flattened tests.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>