refactor: expose `try_add_extension` at Loader and Daemon
Users manually creating a Zenoh-Flow runtime or daemon could need to add
extensions. This commit makes it easier to add extensions by exposing the
function `try_add_extension` both at the Loader and Daemon levels.

The `Extension` structure was also made public with methods to access the values
of the different fields composing it.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>