refactor: improve data flow state management (#211)
The main changes introduced by this commit pertain to the `InstanceState`
structure.

Two variants were added:
- Creating
- Failed

For each variant a timestamp was added, e.g. `Creating(Timestamp)`, such that
users can track when modifications were performed on their data flow.

The `Failed` state had additional consequences, on the runtime, daemon and
zfctl.
The purpose of introducing this state is to allow users to keep track of what
went wrong when trying to create their data flow. In order to do so, we need to
keep the failed instance within the runtime --- we cannot immediately drop it as
was done before.

If we keep the failed instance, we suddenly "allow" start / abort requests on
it. Which should not be done. The changes on the runtime, daemon and zfctl
address this problem.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>