refactor(daemon): selector creation returns no error
Although panicking seems like a strong posture, we believe this choice strikes a
correct balance: we, the Zenoh-Flow team, know the internals of Zenoh and can
guarantee the validity of the key expressions we are building. Plus, all the
exposed API leveraging this function only accepts `RuntimeId` which is a thin
wrapper over ZenohId.

The benefit is that this simplifies our API, removing an error that,
realistically, would never occur.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>