feat(standalone-daemon): accept optional Zenoh/Zenoh-Flow configurations
With this change, users can specify, in a dedicated file, the
configuration of the Zenoh-Flow Runtime as well as Zenoh's.

This allows setting the name and supported extensions in a dedicated
file, passed via the '-c' option.

Similarly to other executable, the configuration for Zenoh can be passed
via the '-z' option.

* zenoh-flow-standalone-daemon/src/main.rs:
  - improve the documentation of the different command line options,
  - make the 'name' option and 'configuration' mutually exclusive (via
    the `group` directive),
  - handle the different options.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>