refactor(daemon): removed feature "plugin"
Because of the way features work in Rust, in particular the fact that
they are additive, it is impossible to have mutually exclusive features.

In the case of the `zenoh-flow-daemon` crate, we wanted to have the
feature "plugin" enabled for the `zenoh-plugin-zenoh-flow` crate and not
for the `zenoh-flow-standalone-daemon`. Having that feature enabled for
one, enabled it for all crates --- which was not the desired behaviour.

To achieve the desired behaviour we should: (i) remove the feature and
(ii) separate the Zenoh configuration from the Zenoh-Flow configuration.

This commit addresses (i).

* justfile: removed the build test with `--features plugin` as this
  feature no longer exists.
* zenoh-flow-daemon/src/daemon/configuration.rs: removed the embedded
  `ZenohConfiguration` structure from the `ZenohFlowConfiguration`.
  We have to separate them.
* zenoh-flow-daemon/src/daemon/mod.rs: removed the code that was called
  with the feature "plugin" disabled,
* zenoh-flow-daemon/src/lib.rs: removed the documentation regarding the
  feature "plugin".
* zenoh-plugin-zenoh-flow/Cargo.toml: removed the feature plugin.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>