feat(nodes): pass the library path in the context
Fixes #246.

As explained in #246, providing the path of the library that Zenoh-Flow
loads is a requirement for developing extensions. For instance, in
Python, the `pyo3` library needs to load the script.

This change adds the accessor method `library_path` to the `Context`
structure, that returns a reference to the full path of the library
being loaded.

* zenoh-flow-nodes/src/context.rs:
  - add the library_path field to the Context structure,
  - update the constructor accordingly,
  - add the accessor method `library_path()`.
* zenoh-flow-runtime/src/loader/mod.rs:
  - update the Loader structure to also store the path of the library
    that is being loaded,
  - update the code to account for the path of the library being also
    stored.
* zenoh-flow-runtime/src/runtime/load.rs:
  - remove the `Context` in the arguments of the `try_load_*` methods as
    it cannot be created at that stage but only when the nodes are
    loaded,
  - create the `Context` where appropriate.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>