fix: in case of extensions, expose path of extension
The previous code was exposing the path of the rust library associated
with the extension, not the path of the library that was "loaded".

This code will load the rust library but keep track and expose the path
of the library that is to load (be it an extension or not).

* zenoh-flow-runtime/src/loader/mod.rs:
  - differentiate the rust library from the "library to load", that, in
    the case of extensions, are different,
  - keep in the hashmap the "library to load" as key.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>