fix(commons): do not fail without `Vars` section
If the `Vars` section was not present in a file that is parsed by the
function `try_parse_from_file` then an error would be thrown.

This commit changes the behaviour of the function such that if it is
missing, no error is returned and instead a log message is recorded.

* Cargo.lock: added dependency `tracing`.
* zenoh-flow-commons/Cargo.toml: added dependency `tracing`.

* zenoh-flow-commons/src/utils.rs: instead of throwing an error if the
  `Vars` section is missing, log a debug message and continue the normal
  execution.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>